requirements-additional.txt
imapclient>=2.2.0
python-dotenv>=0.21.0
python-magic>=0.4.27
Pillow>=9.5.0
psutil>=5.9.0
